Fixing keyboard inheritance work since merge.
